+++ packages/argon-router-core/lib/create-route.ts
@@ -13,7 +13,7 @@ import {
   RouteOpenedPayload,
 } from './types';
 
-import { ParseUrlParams } from 'typed-url-params';
+import { ParseUrlParams } from '@argon-router/paths';
 
 interface Config<T> {
   path: T;
@@ -21,12 +21,34 @@ interface Config<T> {
   beforeOpen?: Effect<void, any, any>[];
 }
 
-type SafeParams<T> = T extends Record<string, never> ? void : T;
-
-export function createRoute<
-  T extends string,
-  Params = SafeParams<ParseUrlParams<T>>,
->(config: Config<T>): Route<Params> {
+/**
+ * @description Creates argon route
+ * @param config Route config
+ * @returns `Route\<Params\>`
+ * @link https://movpushmov.dev/argon-router/core/create-route.html
+ * @example ```ts
+ * import { createRoute } from '@argon-router/core';
+ *
+ * // basic
+ * const route = createRoute({ path: '/route' });
+ * route.open();
+ *
+ * // with params
+ * const postRoute = createRoute({ path: '/post/:id' });
+ * //       ^---  Route<{ id: string }>
+ *
+ * // with parent
+ * const profile = createRoute({ path: '/profile/:id' });
+ *
+ * const friends = createRoute({ path: '/friends', parent: profile });
+ * const posts = createRoute({ path: '/posts', parent: profile });
+ *
+ * posts.open(); // profile.$isOpened -> true, posts.$isOpened -> true
+ * ```
+ */
+export function createRoute<T extends string, Params = ParseUrlParams<T>>(
+  config: Config<T>,
+): Route<Params> {
   let asyncImport: AsyncBundleImport;
 
   type OpenPayload = RouteOpenedPayload<Params>;
@@ -49,7 +71,7 @@ export function createRoute<
       if (parent) {
         await parent.internal.openFx({
           ...(payload ?? { params: {} }),
-          historyIgnore: true,
+          navigate: false,
         });
       }
 
@@ -69,6 +91,8 @@ export function createRoute<
   const openedOnServer = createEvent<OpenPayload>();
   const openedOnClient = createEvent<OpenPayload>();
 
+  const navigated = createEvent<OpenPayload>();
+
   const closed = createEvent();
 
   sample({
@@ -79,7 +103,7 @@ export function createRoute<
   const defaultParams = {} as Params;
 
   sample({
-    clock: openFx.doneData,
+    clock: navigated,
     fn: (payload): Params => {
       if (!payload) {
         return defaultParams;
@@ -91,7 +115,7 @@ export function createRoute<
   });
 
   split({
-    source: openFx.doneData,
+    source: navigated,
     match: () => (typeof window === 'undefined' ? 'server' : 'client'),
     cases: {
       server: openedOnServer,
@@ -106,19 +130,12 @@ export function createRoute<
   });
 
   sample({
-    clock: opened,
-    fn: () => true,
-    target: $isOpened,
-  });
-
-  sample({
     clock: close,
     target: closed,
   });
 
   sample({
-    clock: closed,
-    fn: () => false,
+    clock: [opened.map(() => true), closed.map(() => false)],
     target: $isOpened,
   });
 
@@ -133,11 +150,13 @@ export function createRoute<
     openedOnClient,
     openedOnServer,
 
+    ...config,
+
     internal: {
+      navigated,
       close,
       openFx: openFx as Effect<any, any, any>,
       setAsyncImport: (value: AsyncBundleImport) => (asyncImport = value),
-      ...config,
     },
 
     '@@unitShape': () => ({