+++ packages/argon-router-react/lib/create-route-view.tsx
@@ -1,5 +1,26 @@
 import { CreateRouteViewProps, RouteView } from './types';
 
+/**
+ * @description Creates Route view without async bundle load
+ * @link https://movpushmov.dev/argon-router/react/create-route-view.html
+ * @param props Route view props
+ * @returns RouteView
+ * @example ```ts
+ * import { createRouteView } from '@argon-router/react';
+ * import { routes } from '@shared/routing';
+ * import { MainLayout } from '@layouts';
+ *
+ * function Profile() {
+ *   return <>...</>;
+ * }
+ *
+ * export const ProfileScreen = createRouteView({
+ *   route: routes.profile,
+ *   view: Profile,
+ *   layout: MainLayout,
+ * });
+ * ```
+ */
 export const createRouteView = (props: CreateRouteViewProps): RouteView => {
   return {
     route: props.route,