+++ packages/argon-router-react/lib/create-routes-view.tsx
@@ -9,6 +9,29 @@ interface CreateRoutesViewProps {
   otherwise?: ComponentType;
 }
 
+/**
+ * @description Create routes view which renders current opened route. `Don't forget add <RouterProvider>`!
+ * @param props Routes view config
+ * @link https://movpushmov.dev/argon-router/react/create-routes-view.html
+ * @returns RoutesView
+ * @example ```tsx
+ * import { createRoutesView } from '@argon-router/react';
+ * import { router } from './router';
+ * // feed screen & profile screen must be created with createRouteView!
+ * import { FeedScreen, ProfileScreen } from './screens';
+ *
+ * const RoutesView = createRoutesView({ routes: [FeedScreen, ProfileScreen] });
+ *
+ * // then you can use it like react component:
+ * function App() {
+ *   return (
+ *     <RouterProvider router={router}>
+ *       <RoutesView />
+ *     </RouterProvider>
+ *   );
+ * }
+ * ```
+ */
 export const createRoutesView = (props: CreateRoutesViewProps) => {
   const { routes, otherwise: NotFound } = props;
 
@@ -16,9 +39,7 @@ export const createRoutesView = (props: CreateRoutesViewProps) => {
     const { activeRoutes } = useRouter();
 
     const filtered = activeRoutes.reduce<Route<any>[]>((acc, route) => {
-      return acc.filter(
-        (r) => r !== (route as InternalRoute<any>).internal.parent,
-      );
+      return acc.filter((r) => r !== (route as InternalRoute<any>).parent);
     }, activeRoutes);
 
     const displayedRoute = routes.find(