+++ packages/argon-router-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.1](https://github.com/movpushmov/argon-router/compare/v0.6.0...v0.6.1) (2025-03-26)
+
+**Note:** Version bump only for package @argon-router/react
+
+# [0.6.0](https://github.com/movpushmov/argon-router/compare/v0.5.1...v0.6.0) (2025-03-26)
+
+### Features
+
+- argon paths ([093592a](https://github.com/movpushmov/argon-router/commit/093592a15a34fd4e4dfd6794dafc0f6704797a0d))
+
+## [0.5.1](https://github.com/movpushmov/argon-router/compare/v0.5.0...v0.5.1) (2025-03-22)
+
+### Bug Fixes
+
+- react tests ([65d365b](https://github.com/movpushmov/argon-router/commit/65d365b977e61b10109972735c95c056485a6307))
+
 # [0.5.0](https://github.com/movpushmov/argon-router/compare/v0.4.0...v0.5.0) (2025-02-26)
 
 **Note:** Version bump only for package @argon-router/react